requirements.txt
CommonMark
watchdog
Jinja2==2.10
markdown2==2.3.5
MarkupSafe==1.0
PyYAML==3.12
